Organize reporter components
