+++ client/app/[slug]/page.tsx
@@ -1,13 +1,13 @@
 import { Footer } from "@/components/Footer";
 import { Header } from "@/components/Header";
-import { Reporter } from "@/components/Reporter";
 import { Analysis } from "@/components/report/Analysis";
 import { BackButton } from "@/components/report/BackButton";
 import { ClientContainer } from "@/components/report/ClientContainer";
 import { Overview } from "@/components/report/Overview";
+import { Reporter } from "@/components/reporter/Reporter";
 import type { Meta, Report, Result } from "@/type";
-import { Box, Separator } from "@chakra-ui/react";
 import { ReportVisibility } from "@/type";
+import { Box, Separator } from "@chakra-ui/react";
 import type { Metadata } from "next";
 import { notFound } from "next/navigation";
 import { getApiBaseUrl } from "../utils/api";
